requirements.txt
fastapi
uvicorn
jinja2
pinecone-client
langchain
python-dotenv
